Allow setting named arguments from attributes file
This shuffles things around a bit but the main thrust is to:

* Load template attributes from an optional file

* Create a context class that picks args out by name, as specified in the
  attributes file

The Context class gets most of this done in an ERB-safe way by making a blank
binding via BasicObject.  Templates make a sublcass of Context that defines the
named argument methods.  Context does not duplicate or modify args.